Copy initramfs to EFI partition only when it's already there
Initramfs needs to be on EFI partition only when booting via xen.efi
directly (which is still possible). kernel-install script detects it
based on xen.cfg presence and handles the copying of new versions. It
handles also removing old versions in that case. For existing versions,
do it only when the initramfs was already copied there before.
Otherwise, older version could accumulate since kernel-install wouldn't
remove them it it wasn't the one putting it there in the first place.

Fixes QubesOS/qubes-issues#8526